Fix warnings

+++ scopeprotocols/DeEmbedFilter.cpp
@@ -244,7 +244,7 @@ void DeEmbedFilter::DoRefresh(bool invert, vk::raii::CommandBuffer& cmdBuf, shar
 
 	//Sanity check: if we have noisy or poor quality S-parameter data, group delay might not make sense.
 	//Skip this correction pass in that case.
-	if(llabs(groupdelay_samples) >= npoints)
+	if(static_cast<size_t>(llabs(groupdelay_samples)) >= npoints)
 	{
 		groupdelay_fs = 0;
 		groupdelay_samples = 0;
